SSCS-9069 Move direction due date validation to mid event callback hander
